pi_kid_voice_bot/requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
sounddevice>=0.4.7
scipy>=1.11.0
simpleaudio>=1.0.4
edge-tts>=6.1.12
